add additional test functions
this adds a version of exercise_path_oram that queries consecutive
locations over and over, rather than using the "progressive"
strategy.

the progressive stratey was good initially when we would find bugs
quickly when items are queried again, and it's also good for testing
very large ORAMs in a way that is interesting.

this strategy helps to answer questions posed in issue #12, like,
is it the case that when all items exist in the tree, there is no
space for any dummy blocks, due to our choices of parameters.
In case of such an issue, we would expect this test to fail when
the number of rounds is significantly larger than the size of the
ORAM, because with high probability when an item is mapped
there would be no space on the branch that it selects or in the
stash.