Remove cargo crev from CI
I think this TODO item "have enough reviews" is not going to happen
in the foreseeable future. In the meantime I have no interest in maintaining
this github action, this tool doesn't provide useful information, and
sometimes it breaks and requires fixups. That's not a good tradeoff
for my time or anyone else's.